Consolidate body-reconstruction facts into MemberBodyFacts (#2777)
Merge the two neutral product->harness extractors (ReferencedNamespaceExtractor
and ConstructorBodyFactExtractor) into a single discoverable MemberBodyFacts
class with focused query methods: ReferencedNamespaces(function) and
Constructor(function). One type now answers the set of neutral reconstruction
questions the harness asks about a decompiled member body, and is the natural
home for future body facts (e.g. backing-field member requirements).

The unrelated Pipeline/Facts/*RecoveryFacts (ILoweringFactProvider lowering-
coverage ledger) are a different subsystem and are left untouched. The
ConstructorBodyFacts DTO is retained as Constructor()'s return type.

Pure consolidation: the IR walks are moved verbatim, so extraction is
byte-neutral. Consumers (ReturnToSenderTypePlanner, FidelityCheck) repointed;
the two test classes merged into MemberBodyFactsTests.

Evidence: full slnx build clean; MemberBodyFactsTests 6; RTS oracle suite 224
(Evidence/FixtureCatalog/Prototype/GeneratedFilter) zero churn.

Copilot-Session: d6c5d668-eb01-4f5b-bc22-09e6eade8835
